requirements_minimal.txt
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
authlib==1.2.1
requests==2.31.0
google-auth==2.23.4 